Issue #34: Decode HTML entities before stripping tags
It looks like the original code dealing with HTML entities
may have had a bug... what it was doing wasn't in line with
the apparent intent from the comments, etc.

Since a cleaned-up version looks like what was originally intended,
and since we're aiming for clear readability and searchability
of the text, I'm going to hold off on adding separate _clean
or _orig fields until we see how this pans out.